fix: Remove `flat-cache` dependency
