+++ packages/cspell/package.json
@@ -88,7 +88,6 @@
     "@cspell/cspell-types": "workspace:*",
     "@cspell/dynamic-import": "workspace:*",
     "@cspell/url": "workspace:*",
-    "@types/flat-cache": "^2.0.2",
     "chalk": "^5.4.1",
     "chalk-template": "^1.1.0",
     "commander": "^14.0.0",
@@ -99,7 +98,7 @@
     "cspell-io": "workspace:*",
     "cspell-lib": "workspace:*",
     "fast-json-stable-stringify": "^2.1.0",
-    "flat-cache": "^5.0.0",
+    "flatted": "^3.3.3",
     "semver": "^7.7.2",
     "tinyglobby": "^0.2.14"
   },
